change values of KNX devices and the likes
basically uncommented the lines, which the original author "paresy"
wrote to change the values of the variables of an instance and modified
them slightly in some points.
Dont have any "real" devices, so I cant test it properly :P.
Coultnt find any dummy for it too and its 0:30 in the morning. Wlep, I
can only hope now but am fairly confident due to the only error in the
scripts being "couldnt open socket", which makes sense because I dont
have a real device